Launching 1

+++ app/src/main/java/com/TronAr/Video/Video.java
@@ -1,23 +1,17 @@
 /*
- * Copyright © 2012 Iain Churcher
+ * Copyright 2015 Azmeer Raja
  *
- * Based on GLtron by Andreas Umbach (www.gltron.org)
+ *      Licensed under the Apache License, Version 2.0 (the "License");
+ *      you may not use this file except in compliance with the License.
+ *      You may obtain a copy of the License at
  *
- * This file is part of GL TRON.
+ *          http://www.apache.org/licenses/LICENSE-2.0
  *
- * GL TRON is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * GL TRON is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with GL TRON.  If not, see <http://www.gnu.org/licenses/>.
- *
+ *      Unless required by applicable law or agreed to in writing, software
+ *      distributed under the License is distributed on an "AS IS" BASIS,
+ *      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *      See the License for the specific language governing permissions and
+ *      limitations under the License.
  */
 
 package com.TronAr.Video;
@@ -75,14 +69,14 @@ public class Video {
 		float ratio = _width / _height;
 		//float znear = 0.5f;
 		float znear = 1.0f;
-		float zfar = (float)(GridSize * 6.5f);
+		float zfar = GridSize * 6.5f;
 		//float fov = 120.0f;
 		float fov = 105.0f;
 		
 		gl.glMatrixMode(GL10.GL_PROJECTION);
 		gl.glLoadIdentity();
-		top = (float)Math.tan(fov * Math.PI / 360.0f) * (float)znear;
-		left = (float)(((float)-top)*((float)ratio));
+		top = (float) Math.tan(fov * Math.PI / 360.0f) * znear;
+		left = -top * ratio;
 		gl.glFrustumf(left, -left, -top,  top, znear, zfar);
 		
 		w = Float.floatToIntBits(_width);
